good

+++ src/main/java/telran/spring/service/ProductService.java
@@ -7,7 +7,7 @@ public interface ProductService {
     Product addProduct(Product product);
     List<Product> getAllProducts();
     List<Product> getProductsByCategory(String category);
-    List<Product> getProductsByPrice(int maxPrice);
+    List<Product> getProductsByPrice(double maxPrice);
     Product editProduct(int id, Product product);
     void deleteProduct(int id);
     List<Product> addProducts(List<Product> products);
